// interface/products.interface.ts
// Generated by https://quicktype.io

export interface IGuitar {
  id: number;
  attributes: IGuitarAttributes;
}

export interface IGuitarAttributes {
  nombre: string;
  createdAt: string;
  updatedAt: string;
  publishedAt: string;
  descripcion: string;
  precio: number;
  url: string;
  imagen: Imagen;
}

export interface Imagen {
  data: DataProduct;
}

export interface DataProduct {
  id: number;
  attributes: DataProductAttributes;
}

export interface DataProductAttributes {
  name: string;
  alternativeText: null;
  caption: null;
  width: number;
  height: number;
  formats: Formats;
  hash: string;
  ext: string;
  mime: string;
  size: number;
  url: string;
  previewUrl: null;
  provider: string;
  provider_metadata: ProviderMetadata;
  createdAt: string;
  updatedAt: string;
}

export interface Formats {
  small: Medium;
  medium: Medium;
  thumbnail: Medium;
}

export interface Medium {
  ext: string;
  url: string;
  hash: string;
  mime: string;
  name: string;
  path: null;
  size: number;
  width: number;
  height: number;
  provider_metadata: ProviderMetadata;
}

export interface ProviderMetadata {
  public_id: string;
  resource_type: string;
}